2.Sensor(Vscode)/requirements.txt
# 材料分类系统依赖包
# Material Classification System Requirements

# 核心数据处理
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 机器学习
scikit-learn>=1.0.0

# 数据可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 串口通信（数据采集用）
pyserial>=3.5

# 键盘控制（数据采集用）
keyboard>=0.13.5

# 可选：深度学习（未来扩展）
# torch>=1.10.0
# torchvision>=0.11.0
